src/requirements.txt
pyinstaller
pyinstaller-hooks-contrib
PyYAML
langchain
langchain-openai
pydantic<2
langchain-ollama